Bump to 2.0.1
This would have been 1.5.2, but the number that the app store cares about being strictly monotonically increasing
is actually the CFBundleVersion--also known as the Build Version. Since we thought that the build version got appended to the
Bundle Version String (different field), we had set it to 2 at some point.

From now on, we shall update both the Build Version and Bundle Version String in lockstep, starting at 2.0.1