[schema] Add PresenceManager to update presence info and post presence data
This also introduces a Core Data schema migration, for the newly added
ZUserPresence managed object. We use ZUserPresence objects for maintaining
presence information from users---they are updated when we receive presence
info from the server. Each presence object has a timestamp for when it was last updated

ZUserPresences are unique on [user, client] pairs.