Delay showing keyboard after tapping a username
There is a weird race condition taking place when both switching
StreamViewControllers *and* trying to trigger the compose view on the
active StreamViewController. This is a bit of a hack, but near
as I can tell a better solution requires a major refactoring of how
narrowing works; this feels like the better option for now.