Add custom server url option to login screen
Also fix logged-in resume of app

Adding a domain field to the login view meant that we could no
longer rely on hardcoded API endpoints when resuming a logged-in state (where
the user would not encounter the login view and set the domain variable again).

We now add the domain as a label in the keychain item, and remove the remaining
hardcoded urls. We also stop using domain in our sqlite db filename.