Rip out dtcoretext
This commit is from acrefoot <acrefoot@zulip.com>

When we tried to target iOS 6, we spent a lot of effort to try to get
fast HTML views. UIWebViews are far too slow, and NSAttributedStrings
didn't have an easy way to be set via HTML data. Thus we used the
DTCoreText project and friends. However, it is a lot of work to keep
that dependency up-to-date. Now that we are open-sourcing this app,
and enough people are using iOS 7+, we can try to use the native
NSAttributedString, which has since enabled a way to set via HTML.

- remove ThirdParty/DTCoreText
- Make project target iOS 7
- Fix some compiler warnings
- Transform emoji_map.json to output HTML entities for the emoji codepoints
- Strip trailing whitespace in NSAttributedString rendered from message HTML
  (put there when you have a <p> tag, and so on)

What Works:
- emoji in the emoji_map.json
- rich text rendering
- (new) ability to select text in messages
- Use Zulip font
- CSS rules to get markdown's HTML to render nicely
- Cell Heights are correct (this one took quite a lot of work, unfortunately)

What doesn't yet work:
- links in messages aren't hooked up by the necessary delegates
- I don't see how, but the message composition UI is now broken

What hasn't been tested:
- relative performance (it seems to be fast enough, but I've only tried on the simulator