Proper showing of LoginController
Putting the home page as root viewcontroller for the loginviewcontroller
can lead to user getting to it without login. like going to login with google
and back he will see a back button to it.

This commit creates a standalone Navigation controller for the Login viewcontroller
present it instead of pushing it, and also retouch the affected part.